requirements.txt
streamlit
opencv-python
face-recognition
Pillow
numpy
pandas
gspread
google-auth 